feat: Adds userGroup filter to nonAdmin endpoint to fetch UserTaskRuns.
Adds class to read identity provider's custom properties that are later on used to select which IdP adapter to use (which for now is just Keycloak).

Adds a Bean to expose an instance of CustomIdentityProviderProperties across the whole Spring context.

Adds method to validate if a nonAdmin user is allowed to access resources assigned to a specific userGroup.

Adds Zitadel as part of the IdP vendors enum so that it can be used later on.