chore: Builds docker image and publishes it to local repo
Modifies Docker Compose file to run Keycloak, LittleHorse and the service locally

Changes the way that OIDC configuration is set to a more dynamic one that reads an external file, and lets Spring know through a specific config in application.yml.

Bumps version of library used to handle OIDC and some Spring Security configurations out-of-the-box. Also, updates respective imports.

Adds a dependency that patches an internal vulnerability in spring-addons-starter-oidc.

Updates and reorganizes README.md with instructions to properly run the service locally. Also, it adds some important notes related to Keycloak and LittleHorse configurations.