fix: Adds missing fields to DTOs used to represent a UserTaskRun and its details
Adds a boolean param which is used to allow null responses in utility methods used to fetch proper IdP adapter.

Introduces a new method in IStandardIdentityProviderAdapter to fetch a UserGroup's details

Adds static methods in DTO classes to properly handle the setting of user and userGroup respectively.

Corrects a small styling issue in README.

Adds and updates respective unit tests.