feat(users-management): adds endpoint to fetch managed Users from Identity Provider (#98)
These changes expose a new endpoint through which UTB Backend clients will be able to fetch Users and their respective groups and roles as long as the HTTP request is made by an Admin User. Currently, this is a Keycloak-only feature, but not limited to that, given the fact that the implementation logic is behind an interface used within the UserManagementService class. Also, the respective OpenAPI Spec docs are updated with relevant data about the new endpoint. 

Furthermore, the Keycloak Client library was upgraded to version 26.0.4, and a condition was added to avoid breaking the fetching process of UserTaskRuns when a userGroup is set to empty String instead of null.

All these changes contribute to the initiative around providing Admin users with the chance of managing users and groups through UTB Backend instead of Keycloak's Admin API.