fix: enables an endpoint to claim UserTaskRuns as Admin and implements groups validation when claiming as NonAdmin (#99)
This new endpoint was introduced as a way to allow Admin users to claim tasks while skipping most validations. Also, the groups validation is introduced as a way of allowing Standard users to claim only those UserTaskRuns that are assigned to userGroups that the requester user is a member of.