feat(users-management): adds endpoint to fetch a managed user's data (#108)
This new endpoint was introduced as a way to allow Admin users to fetch a single user's data without having the need to filter out from a list of users.

Also, there is a refactor in client and realm roles fetching to avoid NPE by using Optional objects instead.

These changes close #105